Fix crash in LwIP when VPN is restarted
When the VPN is restarted, in some cases, when the stars align
exactly the way they should, the app goes down with a crash
in what looks like LwIP's ip4_route overstepping its memory
boundaries resulting in SIGSEV/SEGV_MAPERR and SIGBUS/BUS_ADRALN.

In particular, the app hit this crash with some consistency
when the firewall mode switched between blocking to allow
when restarting the vpn. This clued us in that LwIP was probably
acting on a tunnel it shouldn't have been; which led to the current
fix. The app doesn't crash (or rather crash at all) with the
same consistency observed before this hail-mary fix to close
the old vpn-adapter and disconnect the existing tunnel before
creating a new vpn-adapter and connecting to the new tunnel.
Works like a charm.


Former-commit-id: 3efecf652ed7f5649ff24f2f071da2bc2cf164ff