Bug fix for the commit
Bug fix for the commit - 105ce71ce4ad8c08d083517e05be93ee2a50c519

Unamed apps not the same as "Unknown" apps

The requests which are missing the UID's are blocked based on the universal
firewall settings. All the other request which are having UID's but missing
package name are treated as Unnamed(UID).