last updated time against on-device blocklists
The last updated time from the downloaded blocklist will
be shown next to the UPDATE button in the settings screen.

The value of the timestamp is updated from jsonObject returned
from the server.

Need to figure out whether last updated time/last check for update
need to be shown in the settings screen.


Former-commit-id: fd9b0418ea673c77e90cf124df2338107f6b792b