HomeScreen UX change, P2 quantile and other bug fixes
#148 - Introduces Shimmer animation for the Start button in the
Home Screen. Also modified the UX for the DNS and Firewall
button. Now the shimmer animation will be there for both the
start button and RethinkDNS in the Home screen.
RethinkDNS modes selection is removed from the Home screen.
Introduced new bottom sheet in the home screen to configure the
modes.

#157 - DNSCrypt anonymous relays names are modified.
#133, #114 - Now the apps which are not part of app range UID's
the apps which don't have packages(eg., GPS) are now added into
firewall list.
#156 - Latency estimate logic has been modified with P2 quantile
estimator logic with constant time solution.
#153 - The URL for configure rethinkdns plus now contains one more
parameter - timestamp(blocklist downloaded time).
#154 - App and blocklist auto update check frequency is modified
from 24 hours. The check will be happening on every Friday and
Saturday.
#158 - Bug fix on the UX of bottomsheet in the network logs.
#167 - Bug fix, by default app from excluded apps are removed. Now
the apps which are part of excluded list are not refreshed and
reassigning of values are removed from the code.

[Recommended] string in the whitelist is removed.