Cleanup of downloaded files, Quantile Measurement fix
#182 - Clean up the folder which had the old download files.
This was introduced in v053, before that the files downloaded as
part of blocklists are stored in external files dir by the
DownloadManager and moved to canonicalPath.
Now in v053 we are moving the files from external dir to canonical
path. So deleting the old files in the external directory. Introduced
two methods in Utilities.

#180 - Fixed the quantile reset logic. Now the latency calculation
reset will happen when there is a change in the DNS configuration
and for every 1000 requests for now.

#175 - Now check for update will see the source of the app
download and the prompt will be shown to the user based on that.

Issue fix - Fix for DNS changing from CloudFlare security to Default DNS.