App update check and UI modifications
Home screen action buttons are converted to rounded corners with
borders.
Increased the sensitivity of the scroll between view pager.
#175 Changes for the App download - Now the check for the
app download will be based on the download source, when the
download source is set to google play then the listener for the
new update will be initiated else the check for the app update is
fetched from the server.

Update on the HomeScreen title for the wishes.

Included the fade_in and fade_out animation for the UI components.


Former-commit-id: 136862030b3ebe3c062b9e2d6a6ed475932469ad