VPN mode changes, Accessibility service changes
#225 Introduced ConnectivityManager.NetworkCallback() to track down
the changes for the "Block connections without VPN" in Settings->
VPN. The callback will restart the VPN services if necessary and
include/exclude the apps which are part of excluded apps list.

#193 - Now the universal firewall(Block apps not in use) feature
will prompt for the application force stop in case of application
crash/failure in getting the accessibility permission.

Removed the scroll calculation done for the swipe from one
fragment to other in the view pager(firewall and DNS screen).
Instead introduced new NestedScrollableHost(FrameLayout) to
detect the swipe.


Former-commit-id: b0be2660e0a0f37afadf25955ca2c071ac983db3