UI - new DNS and Firewall card in home screen
The card and the button for the DNS and Firewall is removed from
the Home screen and introduced the card which will have the
details of the DNS and Firewall with new CONFIGURE button.

The new UI will take the Firewall screen all apps screen when the
configure is clicked else will be going to the network logs screen.

For the DNS, the configure button will take to the DNS configure
screen and the card click will take to the DNS logs screen in the
DNS view pager.