Application theme implementation - first pass
Introducing the application theme - initial commit.

New settings introduced in the Settings->General->Theme.
Now user can either chose dark mode or light mode for the
application.

The initial commit will have all the attributes set for different
themes. As of now, the light theme is not completely configured.

The theme color are placed in styles.xml which can be configured
to change the colors of both light and dark mode.

Moved some of the string literals to strings.xml.