Application themes - Second pass, Download bug fix
Implementation of adaptive theme for the application. Introduced
new settings in Settings fragment where the user can select the
theme(adaptive, light and dark). The changes for the background
and labels are made accordingly.
https://github.com/celzero/rethink-app/issues/230

Download fix- The download manager for the blocklist download was having an issue downloading the files in batch, Some cases
the download fails in the half way. Changes are made so that the
files are downloaded completely or prompt user about the error.

https://github.com/celzero/rethink-app/issues/220