Adding custom DoH URL
The logic of adding the custom DoH URL to the configure DNS
screen is modified, earlier the URL validation check was sent to
Go code to validate. Once validated then the URL is added.

Now, the custom URL will be added to Configure DNS screen.
The validation and error will be observed once the URL is
connected. (basic DNS url validation is performed during adding
custom DNS)