Fix on some of the build output warnings
Some of the build output warnings are fixed. More fixes to come
for the same.

Removed tun2socks folder - Now the Tun2socks aar file is
generated using firestack. Old files which are part of old version
of the build is removed.

The unused context and unused variables in the listeners are
removed.