Build flavors instead of download source
https://github.com/celzero/rethink-app/issues/210

Check for application update was using download source - play
update and others. Now post fdroid release the app update
check should happen for apps which are downloaded from
playstore and website. So the the download source checks are
modified to build flavor check.
https://f-droid.org/en/docs/Release_Channels_and_Signing_Keys/

PLAY, FDROID and WEBSITE flavors.
Local blocklist will be available for the website and fdroid flavors.

application dependencies are updated to the suggested latest
version.