Fix for apps losing connectivity
https://github.com/celzero/rethink-app/issues/248

Apps lose connectivity during switch over Mobile Data from WiFi.

Internet connection showing metered even if its not the case.

Earlier implementation, the restart of VPN service happened whenever there is network change.

With the new implementation, frequency of restarting the VPN is
reduced which caused the issue, the below fix will add the
network received from the onAvailable() to
setUnderlyingNetworks if its not already part of it.

The active network will be added as first in the array.