Local blocklist download on lockdown mode
When the lockdown mode is enabled for the RethinkVPN service.
Local blocklist were failing to download.

Now, a dialog will be shown if the local blocklist download is
initiated when the lockdown mode is enabled.

Also added timeout for the download, now if the download
is taken mode than 40*1000*60 then the download will be
considered as failed and error will be thrown by the work manager.

The constant values - Constants.WORK_MANAGER_TIMEOUT


Former-commit-id: 41fbcf5a01f74a201317378c14c688eac8481ca2