Introduce UID cache
Cache the UID for the next 30 seconds. In some cases, when the
for the same request the UID returned from connection manager
(getConnectionOwnerUid) is 0 which will reflect in the network
logs as Android.

Now the cache will store the UID's of apps which made connection
in last 30 seconds. UID are stored in the cache(key) with the
combination of protocol, local, remote.