Remove sentry - Commit error
Build.gradle file is having the implementation of io.sentry which
the application is not using anywhere. It inherited from Intra.

F-Droid flagged the app as the sentry is Application
Monitoring and Error Tracking Software.

Removed the implementation from build.gradle.

#258

Added wrong files with the commit. Revert and adding as new commit.