Waiting state fix
#57

When the application crashes/updates it goes into
red waiting state. This causes confusion to the users also requires
click of START button twice to start the app.
FIX : The check for the controller state. If the state of the controller
is activationRequested and the VPN is not connected then
the start will be initiated.