Download manager - File download error
#262 - App crashes when the local blocklist is enabled in
SettingsFragment
Some cases, the DownloadManager returns the status as
STATUS_SUCCESSFUL
even if the file size is 0. The previous check
was dependent on the download manager status.

New logic for the download validation - Post the check for number
of files downloaded validation, validate the downloaded files by
create localBraveDNS object. If the object is returned by the
Dnsx then valid. Null/exception will be invalid.

Future validation - checksum.


Former-commit-id: beb2948885f3ee2d4b0dead7a894eb390880bf7f