Remove startVPN(), fix for firewall rules
Remove the startVPN() usage in BraveVPNService.

Whenever the VPN is started, the startVPN will be
executed once the ConnectivityManager get the network
available callback.

The call for startVPN() is removed. Instead, In
onStartCommand() of BraveVPNService invoke the
restartVPN() to start the VPN.

https://github.com/celzero/rethink-app/issues/90
https://github.com/celzero/rethink-app/issues/242

Still the restartVPN thread is holding up thread.
Figure out a solution for that.

https://github.com/celzero/rethink-app/issues/261

When auto-start or always-on enabled initial requests
processed before loading of firewall rules.

In BraveVPNService, the loadAppFirewallRules() will load
the firewall rules before allowing any connections.

The exponential calculation of wait time until
we load the firewall rules. The other check of the block()
method will be executed once we load the firewall rules.


Former-commit-id: 80858a53ea6086de140c46b7dd8cebe260d297ce