Fetch favico during DNS transaction
Downloads the favicon of the DNS requests during the transaction
and store in the Glide cache. The fav icon will later be
retrieved by the DNS query list / the bottom sheet in the DNS log screen.

Initial implementation was to download the favicon when the
user scrolls the DNS logs / opens the bottom sheet which resulted
in more number of network calls. To avoid the unnecessary calls,
the cache mechanism is introduced.