Connection monitor changes
Setunderlying network logic of v053e is modified.

Now new settings introduced, Set all available networks to the
VPN. If the setting is enabled, the underlying network will be
set with all the available networks. With active network
as first in the list.

If the settings is disabled, only active network will be set in the
underlying network option.

Single threaded handler is implemented to process the messages
from the connectivitiy-manager.