VpnService.Builder.establish() sporadically null
According to the documentation, the establish function
will return null if the VpnService has not been prepared,
however we are using prepare() prior to establish().

Some cases, the ParcelFileDescriptor is null. This results
the app in the red waiting state during the vpn
restart. Fix will stop the vpn complete in case if the
builder.establish() returned as null.

Move the code - open the vpn profile to a common
place. (called from both Home screen and about
page).