Always on issue fix for android version 12 and above
#325

From version 12+(R) android started annotating the
Settings.Secure constants (@hide/@show).

Those constants which are annotated as @hide will
only be available for system_server or system apps
alone.

The constant "always_on_vpn_app" is marked as
@hide.

As part of fix, introduced new method in Utilities which
will check for the alwaysOn based on version.

For versions above 29(Q), there is vpnService.isAlwaysOn
property to check whether always-on is enabled.

For versions prior to 29 the check is made
with Settings.Secure.