Missing download files fix
Bug: When files from canonical path is deleted
but latest timestamp is updated in persistent
state then the download is not initiated.

Fix: the blocklist download will be
initiated when the files are not available
in the application canonical path or when
there is no update in persistent state.

Similar changes done for the remote blocklists as
well.

Renamed the method name from
deleteUnwantedLocalBlocklistFolders to
cleanupOldLocalBlocklistFolders.