Backup and restore
New feature: Backup and Restore
The Rethink based settings stored in both the Shared Preference and database are
backed up in to a separate folder so that user can able to restore those settings
again.

Limitations: As of now, backup and restore works with current app version.
version details are not handled with this changes.