purge older logs
Purge dns and firewall logs.

Earlier the implementation is to retain only configured number of rows
in the database for both dns and firewall logs (configured as 10k).

Now modified the implementation to retain the logs for 7 days.

purge logs older than 7 days (on version v053l, subject to change in later
versions based on user configuration)