new firewall status: bypass DNS and Firewall
This commit adds a new feature that allows application network traffic to bypass both DNS and the firewall. This
can be done by applying a bulk bypass DNS and firewall
rule from the app list, changing the app's firewall status
from the network log's bottom sheet, or applying it from
the app-info screen.