new features: underlying networks, onQuery checks, etc.,
Network reachability checks: When accessing the list of
available or active networks, make a note of the
isReachable() value for each network. This can be done by
performing connectivity checks on both IPv4 and IPv6
addresses based on the link address of the network. This
information is used to add the network to the VPN builder.
Only add a v4 route if the network contains a reachable IPv4
link address, and similarly for IPv6.

Response during onQuery(): During the onQuery() call, the
default response was set as Dns.Alg for DNS+Firewall mode.
Instead, check if the given domain is available in any of the
app's or global trusted lists, or if any app is set to bypass
DNS+Firewall. If none of these conditions are satisfied,
respond to the onQuery() with Preferred instead of Alg.

base uid: Retrieve the app ID (base UID) associated with a
given UID, removing the user ID component from it before
recording the app's entry in network logs.