Issue fix:  #875
The issue is R8 obfuscation throws a ClassCastException when a custom TypeAdapterFactory is used with Retrofit. The issue is caused by the obfuscation process renaming the classes used in the TypeAdapterFactory, which causes the ClassCastException. The obfuscation process have affected the class names on release builds.

As suggested by the issue(square/retrofit#3751), using proguard rule to keep the class names unchanged during obfuscation will solve the issue. Below are the rules added in proguard-rules.pro