Fix: In Logs, show all apps with the same uid #880
Now, there is no way to display the app list in the network
log if there are more than one app. When you click on the
app name in the bottom sheet of the network log, it will
navigate to the app info screen. The 'i' icon on the screen
will now show the list of app names.