Fix #999 Greyed-out network settings
The network settings `enable network visibility` and `do not
route private IP` are visually disabled, but the user can still
change the settings, which have no effect in lockdown
mode. This behavior remains consistent with the previous
version. To avoid confusion, the option for
enabling/disabling the settings will be disabled during
lockdown mode.