Fix: #1010 Vector drawable issue in Android 6
When the apps screen and logs screen are in the process of
loading, an exception is being thrown due to the default
icon that is displayed when an app icon is missing. This
exception states that the required resource cannot be
located. It appears that the Android support library is unable
to identify the vector graphic that includes the <aapt:attr>
attribute. The problem was resolved by substituting the
problematic icon with a different one. Interestingly, this
problem did not arise when using Android Gradle version 7
or earlier, especially when the desugaring feature was
enabled.