Fix: #1006 padding to bottom of recycler
When the RecyclerView contains many entries, those near
the bottom (close to FAB) become hard to access. We could
add padding to the RecyclerView to fix this and enhance
user interaction. This padding will create a buffer space
between the last entries and the FAB.

Adding 100dp as padding bottom