fx: #1915:  remove delay in scroll state change
A delay was previously added to reduce flickering during
scroll state changes, but it didn’t noticeably improvement and  instead caused a crash (IllegalStateException) when the
fragment’s view was no longer available.

To avoid lifecycle-related issues, the delay has been
removed. If needed in the future, we can reintroduce it
using lifecycleScope.